Auto add prefix to userinfo.email and userinfo.profile (#90)
Auto add prefix to userinfo.email and userinfo.profile

This fixes the issue mentioned in #89 . "userinfo.email" and "userinfo.profile" needs to have the full prefix attached, unlike "email" and "profile". The fix is to update the openid regex introduce in #71 to use exact match, so it doesn't accidentally match the userinfo scopes when omitting prefix.