Add support for UAT.
Added "--uat" boolean flag to perform UAT exchange, converting the
fetched OAuth Access Token to UAT. This flag is designed to be
compatible with oauth and sso, as well as curl and caching.

Added "--user_project" GUAC parameter for use in UAT claim.

Refactored and simplified main.go and tasks.go by introducing
the concept of "taskSettings".

Refactored SSO code path to align with oauth and jwt code path,
so it can take advantage of caching and uat support.

Updated tests.