Add support for STS (#92)
Add support for STS.

Added "--sts" boolean flag to perform STS exchange, which converts the
fetched Access Token to STS token. This flag is designed to be
compatible with oauth and sso, as well as curl and caching.

Added "--quota_project" GUAC parameter for use in STS claim.

Refactored and simplified main.go and tasks.go by introducing
the concept of "taskSettings".

Refactored SSO code path to align with oauth and jwt code path,
so it can take advantage of caching and STS support.

Updated tests.